Create lineparser.py
This was written with Python 3.4.1
Double click the python script while it's in the same file location as sample_data.txt

A quick and dirty python script that runs through the text line by line, then separates each comma delimited string from each line checking to make sure that the radius value is in meters instead of kilometers. The unit of measure check and conversion should probably be put in to a function.